+++ terranet/config.php
@@ -1,0 +1,8 @@
+<?php
+define('PSM_DB_HOST', 'localhost');
+define('PSM_DB_PORT', '3306');
+define('PSM_DB_NAME', 'mon');
+define('PSM_DB_USER', 'qe3');
+define('PSM_DB_PASS', 'J4r4nG0yang');
+define('PSM_DB_PREFIX', 'psm_');
+define('PSM_BASE_URL', 'https://nms.terranet.id');