+++ terranet/src/lang/es_ES.lang.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * TerraNet Monitoring
+ * PHP TERRANET MONITORING
  * Monitor your servers and websites.
  *
- * This file is part of TerraNet Monitoring.
- * TerraNet Monitoring is free software: you can redistribute it and/or modify
+ * This file is part of PHP TERRANET MONITORING.
+ * PHP TERRANET MONITORING is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * TerraNet Monitoring is distributed in the hope that it will be useful,
+ * PHP TERRANET MONITORING is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with TerraNet Monitoring.  If not, see <http://www.gnu.org/licenses/>.
+ * along with PHP TERRANET MONITORING.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @package     phpservermon
  * @author      Klemens Häckel <http://clickdimension.wordpress.com/>
@@ -40,7 +40,7 @@ $sm_lang = array(
     'locale_tag' => 'es',
     'locale_dir' => 'ltr',
     'system' => array(
-        'title' => 'Server Monitor',
+        'title' => 'TERRANET MONITORING',
         'install' => 'Instalar',
         'action' => 'Acción',
         'save' => 'Guardar',
@@ -451,7 +451,7 @@ $sm_lang = array(
         'password_repeat' => 'Repita la contraseña',
         'password_forgot' => '¿Has olvidado tu contraseña?',
         'password_reset' => 'Perdí mi contraseña',
-        'password_reset_email_subject' => 'Restablecer la contraseña para TerraNet Monitoring',
+        'password_reset_email_subject' => 'Restablecer la contraseña para PHP TERRANET MONITORING',
         'password_reset_email_body' => 'Utilice el siguiente enlace para restablecer tu contraseña. Tenga en cuenta
  que expira de 1 hora.<br><br>%link%',
         'error_user_incorrect' => 'El nombre de usuario proporcionado no se pudo encontrar.',