+++ terranet/src/lang/fi_FI.lang.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * TerraNet Monitoring
+ * PHP TERRANET MONITORING
  * Monitor your servers and websites.
  *
- * This file is part of TerraNet Monitoring.
- * TerraNet Monitoring is free software: you can redistribute it and/or modify
+ * This file is part of PHP TERRANET MONITORING.
+ * PHP TERRANET MONITORING is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * TerraNet Monitoring is distributed in the hope that it will be useful,
+ * PHP TERRANET MONITORING is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with TerraNet Monitoring.  If not, see <http://www.gnu.org/licenses/>.
+ * along with PHP TERRANET MONITORING.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @package     phpservermon
  * @author      Sami Nieminen <nieminen.sami2@gmail.com>
@@ -37,7 +37,7 @@ $sm_lang = array(
     'locale_tag' => 'fi',
     'locale_dir' => 'ltr',
     'system' => array(
-        'title' => 'Server Monitor',
+        'title' => 'TERRANET MONITORING',
         'install' => 'Asenna',
         'action' => 'Toiminta',
         'save' => 'Tallenna',
@@ -282,7 +282,7 @@ $sm_lang = array(
         'password_repeat' => 'Salasana uudestaan',
         'password_forgot' => 'Unohtuiko salasanasi?',
         'password_reset' => 'Resetoi salasanasi',
-        'password_reset_email_subject' => 'Resetoi salasanasi TerraNet Monitoringiin',
+        'password_reset_email_subject' => 'Resetoi salasanasi PHP TERRANET MONITORINGiin',
         'password_reset_email_body' => 'Ole hyvä ja käytä seuraavaa linkkiä restoidaksesi salasanasi. Huomaa että
  linkki vanhentuu tunnin sisällä.<br><br>%link%',
         'error_user_incorrect' => 'Annettua käyttäjänimeä ei löydetty.',