+++ terranet/src/lang/sv_SE.lang.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * TerraNet Monitoring
+ * PHP TERRANET MONITORING
  * Monitor your servers and websites.
  *
- * This file is part of TerraNet Monitoring.
- * TerraNet Monitoring is free software: you can redistribute it and/or modify
+ * This file is part of PHP TERRANET MONITORING.
+ * PHP TERRANET MONITORING is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * TerraNet Monitoring is distributed in the hope that it will be useful,
+ * PHP TERRANET MONITORING is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with TerraNet Monitoring.  If not, see <http://www.gnu.org/licenses/>.
+ * along with PHP TERRANET MONITORING.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @package     phpservermon
  * @author      andlil
@@ -37,7 +37,7 @@ $sm_lang = array(
     'locale_tag' => 'sv',
     'locale_dir' => 'ltr',
     'system' => array(
-        'title' => 'Server Monitor',
+        'title' => 'TERRANET MONITORING',
         'install' => 'Installera',
         'action' => 'Åtgärd',
         'save' => 'Spara',
@@ -299,7 +299,7 @@ $sm_lang = array(
         'password_repeat' => 'Upprepa lösenord',
         'password_forgot' => 'Glömt lösenord?',
         'password_reset' => 'Återställ lösenord',
-        'password_reset_email_subject' => 'Password reset for TerraNet Monitoring',
+        'password_reset_email_subject' => 'Password reset for PHP TERRANET MONITORING',
         'password_reset_email_body' => 'Anv&auml;nd f&ouml;ljande l&auml;nk f&ouml;r att &aring;terst&auml;lla ditt
  l&ouml;senord. T&auml;nk p&aring; att l&auml;nken bara &auml;r giltig 1
  timme.<br><br>%link%',