+++ terranet/src/lang/uk_UA.lang.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * TerraNet Monitoring
+ * PHP TERRANET MONITORING
  * Monitor your servers and websites.
  *
- * This file is part of TerraNet Monitoring.
- * TerraNet Monitoring is free software: you can redistribute it and/or modify
+ * This file is part of PHP TERRANET MONITORING.
+ * PHP TERRANET MONITORING is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * TerraNet Monitoring is distributed in the hope that it will be useful,
+ * PHP TERRANET MONITORING is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with TerraNet Monitoring.  If not, see <http://www.gnu.org/licenses/>.
+ * along with PHP TERRANET MONITORING.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @package     phpservermon
  * @author      Oleksa Vyshnivsky <dying.escape@gmail.com>
@@ -36,7 +36,7 @@ $sm_lang = array(
     'locale_tag' => 'uk',
     'locale_dir' => 'ltr',
     'system' => array(
-        'title' => 'Server Monitor',
+        'title' => 'TERRANET MONITORING',
         'install' => 'Встановити',
         'action' => 'Дія',
         'save' => 'Зберегти',
@@ -514,7 +514,7 @@ $sm_lang = array(
         'password_repeat' => 'Повторити пароль',
         'password_forgot' => 'Забули пароль?',
         'password_reset' => 'Скинути пароль',
-        'password_reset_email_subject' => 'Скинути свій пароль до TerraNet Monitoring',
+        'password_reset_email_subject' => 'Скинути свій пароль до PHP TERRANET MONITORING',
         'password_reset_email_body' => 'Будь ласка, скористайтеся наступним
  посиланням для скидання свого пароля. Будь
  ласка, пам’ятайте, що воно діє протягом 1